Editor: gave proper names to StartNewGame wizard page classes
